+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic versioning](http://semver.org/).
 - do not flip arguments on `ListL/ConcatenationMonoid`
 - introduce `Alt` and `Plus` typeclasses
 - define `ConstantSemigroup`
-- fix `foldr` behavior for lists
+- correct `foldr` implementation for `ListL` and `Traversable`
 
 ## [2.0.0] - 2023-03-21
 
@@ -22,6 +22,7 @@ and this project adheres to [Semantic versioning](http://semver.org/).
 
 - introduce `OppositeSemigroup` and `OppositMonoid`
 - do not flip arguments on `ListL/ConcatenationMonoid`
+- correct `foldr` implementation for `ListL` and `Traversable`
 
 ## [1.3.0] - 2023-08-23